// src/zustand/types.tsx
export interface Post {
	userId: number
	id: number
	title: string
	body: string
}

export interface AppState {
	users: User[]
	userDetail: User | null
	posts: Post[]
	isLoading: boolean
	getUsers: () => void
	getUserDetail: (userId: string | number) => void
	getUserPosts: (userId: string | number) => void
}

export interface User {
	id: number
	name: string
	username: string
	email: string
	address: Address
	phone: string
	website: string
	company: Company
}

export interface Address {
	street: string
	suite: string
	city: string
	zipcode: string
	geo: Geo
}

export interface Geo {
	lat: string
	lng: string
}

export interface Company {
	name: string
	catchPhrase: string
	bs: string
}